Scalar: Updates and convert tests to Catch.
